requirements.txt
flask
pydantic-settings
pdf2image
werkzeug
requests
streamlit
pandas
google-generativeai
transformers
torch
sentencepiece
accelerate
Pillow
gunicorn